Fix a memory bug and move predict vector computation into tree class

+++ include/falcon/model/model_io.h
@@ -34,8 +34,8 @@ void load_lr_model(const std::string& model_save_path, int& weight_size, Encoded
  * @param lr_testing_accuracy: model testing accuracy
  * @param report_save_path: report file to be saved
  */
-void save_lr_report(float lr_training_accuracy,
-    float lr_testing_accuracy,
+void save_lr_report(double lr_training_accuracy,
+    double lr_testing_accuracy,
     const std::string& report_save_path);
